function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
psycopg2-binary==2.8.5
sendgrid==6.3.0

